Don't "check" author inherent
We never needed this check to begin with.
I wrote it when I didn't properly understand how/why inherents were checked.
I left it in because it was harmless.
Now that we have a proper slot beacon,
the checking is incompatible with the relay chain beacon.
So it finally gets pruned.